+++ lib/src/components/components.dart
@@ -14,6 +14,7 @@ export 'combobox/foss_combobox.dart';
 export 'date_picker/foss_date_picker.dart';
 export 'dialog/foss_dialog.dart';
 export 'drawer/foss_drawer.dart';
+export 'list_tile/foss_list_tile.dart';
 export 'meter/foss_meter.dart';
 export 'number_field/foss_number_field.dart';
 export 'otp_field/foss_otp_field.dart';